+++ ASCOfficePPTFile/PPTFormatLib/Records/SlideProgTagsContainer.h
@@ -371,7 +371,7 @@ public:
         LONG lCurLen(0);
         SRecordHeader ReadHeader;
 
-        while (lCurLen < m_oHeader.RecLen)
+        while (lCurLen < (LONG)m_oHeader.RecLen)
         {
             if ( ReadHeader.ReadFromStream(pStream) == false)
                 break;
