+++ ASCOfficePPTFile/PPTFormatLib/Records/StyleTextProp11Atom.h
@@ -56,7 +56,7 @@ public:
         LONG lCurPos		=	0;
         StreamUtils::StreamPosition ( lCurPos, pStream );
 
-        while ( lPos + m_oHeader.RecLen > lCurPos)
+        while ( lPos + (LONG)m_oHeader.RecLen > lCurPos)
         {
             SStyleTextProp11 style;
             style.ReadFromStream(pStream);
